Release 0.99a.
Major changes : 
- Physics now uses a single thread. This simplification greatly improves code readability.
- Physics has now meter coordinates instead of pixel coordinates. This works much better but this might have broken some old cold because almost everything has been rewritten.
- New debug renderer for physics
- Physics uses float everywhere
- New font generation method that allows scaling (see DemoFontGeneration)
- Graphics now allows to change alpha of pictures drawn (alpha demo)

Minor changes
- Some new demos (alpha blending, simple light, physics...)
- Some demos rewritten more nicely (DemoPhysicsBall for instance)
- Data folder has been reorganized
- Physics simulator update methods

+++ trunk/gdx2d-desktop/src/hevs/gdx2d/components/physics/PhysicsCircle.java
@@ -1,5 +1,7 @@
 package hevs.gdx2d.components.physics;
 
+import hevs.gdx2d.lib.physics.AbstractPhysicsObject;
+
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.physics.box2d.Shape.Type;
 
@@ -10,11 +12,11 @@ import com.badlogic.gdx.physics.box2d.Shape.Type;
  */
 public class PhysicsCircle extends AbstractPhysicsObject{
 	
-	public PhysicsCircle(String name, Vector2 position, int radius) {
+	public PhysicsCircle(String name, Vector2 position, float radius) {
 		super(Type.Circle, name, position, radius, radius, 10f, 0.6f, 0.6f, true);			
 	}
 	
-	public PhysicsCircle(String name, Vector2 position, int radius, float density, float restitution, float friction) {
+	public PhysicsCircle(String name, Vector2 position, float radius, float density, float restitution, float friction) {
 		super(Type.Circle, name, position, radius, radius, density, restitution, friction, true);					
 	}
 }
